feat: adiciona resolução da Lista 03 de exercícios em Java
- Implementação de algoritmos com estruturas condicionais (if/else)
- Uso de switch-case para menus e operações
- Aplicação de operador ternário em diferentes cenários
- Cálculos envolvendo IMC, descontos, impostos e validações
- Exercícios de lógica: divisibilidade, múltiplos, triângulos e intervalos
- Simulações práticas: login, calculadora e menu interativo

Objetivo: praticar lógica de programação e fundamentos da linguagem Java

+++ 1-semestre/java-projects/prjLista03/src/pctExc3/Exercicio03.java
@@ -6,7 +6,7 @@ public class Exercicio03 {
         Scanner sc = new Scanner(System.in);
 
         System.out.println("Digite abaixo seu Saldo Médio para verificação de possível Crédito Especial aplicado:");
-        double saldoMedio = sc.nextInt();
+        double saldoMedio = sc.nextDouble();
 
         if (saldoMedio<=500){
             System.out.println("Seu saldo médio foi de " + saldoMedio + "e não houve crédito especial");
